Third version of the image without VOLUMEs.
Remove VOLUMEs from Dockerfile since they aren't necessary to use
volumes.  However, do keep the directory variables so that they can be
used to archive the data and restore it from the start script when the
volume is mounted.

Also, add the logs volume to the build script and docker-compose.yml
file.

Clean up the create-container script.  Add support for the logs volume
and clean up verbiage.

See: #2