Improve build checks to detect when to build.
Use checksums to determine when files have changed since the last build
so that the image can be rebuilt.  Timestamps don't work well when the
timestamp changes with branch changes that remove/add the files in
between builds.  Also, check the against the latest published base image
to see if that changed.

See: #83